requirements.txt
requests
pillow
torch
numpy
fastapi
pydantic
transformers
uvicorn